% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download MODIS snow data}
\usage{
download_data(date, sat = "MYD10A1", h = 10, v = 10, ...)
}
\arguments{
\item{date}{Day for which snow data should be downloaded as POSIXct.}

\item{sat}{Satelite mission used. Currently only "MYD10A1" is supported.}

\item{h}{horizontal tile number. See also details.}

\item{v}{vertical tile number. See also details.}

\item{...}{further arguments passed to get_tile().}
}
\value{
RasterLayer.
}
\description{
Downloads MODIS snow data and returns a tile as raster tile.
}

