% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_tile}
\alias{get_tile}
\title{Download a MODIS tile}
\usage{
get_tile(ftp, tile, progress = FALSE, clean = TRUE)
}
\arguments{
\item{ftp}{Address of the repository.}

\item{tile}{Name of the tile.}

\item{progress}{Indicates whether or not progress is displayed.}

\item{clean}{Indidcates whether or not temporary files are deleted.}
}
\value{
RasterLayer
}
\description{
Download a MODIS tile
}

